fix(arch): make lib/hive/errors.rb authoritative for the error taxonomy
The dedicated clean-loadable errors boundary owned only the exit-code
contract and the two base classes (Hive::Error, Hive::ConfigError),
while lib/hive.rb re-declared the entire concrete taxonomy — nearly
400 lines of cross-feature error classes (InvalidTaskPath,
ConcurrentRunError, ProviderRouteFailed, WrongStage, …) inside the
root entrypoint. That mirrored the pre-fix Schemas layout, where the
root entrypoint owned the schema namespace until it was moved into
lib/hive/schemas.rb.

Move the concrete taxonomy verbatim into lib/hive/errors.rb so the
dedicated file is the single authority for both the exit-code contract
and the error classes. lib/hive.rb keeps only require_relative
"hive/errors"; every `require "hive"` consumer sees the same public
constant paths, and exit-code overrides plus inheritance (IS-A for
exit-code convenience, e.g. ProviderRouteFailed < AgentError) are
unchanged.

Add test/unit/errors_authority_test.rb to pin the ownership contract:
the taxonomy lives in errors.rb, the root entrypoint declares no error
subclasses, and the file is loaded exactly once per require.
